Refactor: move `mapSearchUserToSupportModel` to fix circular dependency
Test Suites: 135 passed, 135 total
Tests:       766 passed, 766 total
Snapshots:   0 total
Time:        4.741 s
