+++ src/app/users/postDeleteOrganisation.js
@@ -15,7 +15,7 @@ const { updateIndex, getById } = require("./../../infrastructure/search");
 const { isSupportEmailNotificationAllowed } = require("../services/utils");
 const {
   getSearchDetailsForUserById,
-} = require("./searchApiHelpers/getSearchDetailsForUserById");
+} = require("./userSearchHelpers/getSearchDetailsForUserById");
 
 const deleteInvitationOrg = async (uid, req) => {
   const invitationId = uid.substr(4);
