+++ src/app/users/utils.js
@@ -28,7 +28,7 @@ const config = require("./../../infrastructure/config");
 const sortBy = require("lodash/sortBy");
 const {
   getSearchDetailsForUserById,
-} = require("./searchApiHelpers/getSearchDetailsForUserById");
+} = require("./userSearchHelpers/getSearchDetailsForUserById");
 
 const delay = async (milliseconds) => {
   return new Promise((resolve) => {
@@ -513,26 +513,6 @@ const callServiceToUserFunc = async (
   }
 };
 
-const mapSearchUserToSupportModel = (user) => {
-  return {
-    id: user.id,
-    name: `${user.firstName} ${user.lastName}`,
-    firstName: user.firstName,
-    lastName: user.lastName,
-    email: user.email,
-    organisation: user.primaryOrganisation
-      ? {
-          name: user.primaryOrganisation,
-        }
-      : null,
-    organisations: user.organisations,
-    lastLogin: user.lastLogin ? new Date(user.lastLogin) : null,
-    successfulLoginsInPast12Months: user.numberOfSuccessfulLoginsInPast12Months,
-    status: mapUserStatus(user.statusId, user.statusLastChangedOn),
-    pendingEmail: user.pendingEmail,
-  };
-};
-
 module.exports = {
   search,
   searchForBulkUsersPage,
@@ -547,5 +527,4 @@ module.exports = {
   removeAllServicesForUser,
   removeAllServicesForInvitedUser,
   callServiceToUserFunc,
-  mapSearchUserToSupportModel,
 };
