+++ test/app/users/getConfirmAssociateOrganisation.test.js
@@ -9,7 +9,7 @@ jest.mock("./../../../src/infrastructure/search");
 jest.mock("login.dfe.api-client/organisations");
 jest.mock("login.dfe.api-client/invitations");
 jest.mock(
-  "../../../src/app/users/searchApiHelpers/getSearchDetailsForUserById",
+  "../../../src/app/users/userSearchHelpers/getSearchDetailsForUserById",
 );
 
 const { getRequestMock, getResponseMock } = require("./../../utils");
@@ -30,7 +30,7 @@ const {
 } = require("login.dfe.api-client/invitations");
 const {
   getSearchDetailsForUserById,
-} = require("../../../src/app/users/searchApiHelpers/getSearchDetailsForUserById");
+} = require("../../../src/app/users/userSearchHelpers/getSearchDetailsForUserById");
 
 const res = getResponseMock();
