+++ test/app/users/utils.getUserDetails.test.js
@@ -6,7 +6,7 @@ jest.mock("./../../../src/infrastructure/search");
 jest.mock("login.dfe.api-client/users");
 jest.mock("login.dfe.api-client/services");
 jest.mock(
-  "../../../src/app/users/searchApiHelpers/getSearchDetailsForUserById",
+  "../../../src/app/users/userSearchHelpers/getSearchDetailsForUserById",
 );
 
 const { getUserRaw } = require("login.dfe.api-client/users");
@@ -14,7 +14,7 @@ const { getServiceRaw } = require("login.dfe.api-client/services");
 const { getUserDetails } = require("./../../../src/app/users/utils");
 const {
   getSearchDetailsForUserById,
-} = require("../../../src/app/users/searchApiHelpers/getSearchDetailsForUserById");
+} = require("../../../src/app/users/userSearchHelpers/getSearchDetailsForUserById");
 
 describe("When getting user details", () => {
   let req;
